Cart Widget Enhancement
Checks if the current post->ID is on the cart and passes the position
via wp script localization to the widget. This way when a download gets
removed from the cart and ajax is enabled the page will reload if the
current location is the single view of that download.

Introduces edd_get_item_position_in_cart();