'Settings Updated' query fix
Current tab navigation does not take into account the `settings-updated` query attached to the URL after saving the settings. This removes that portion of the URL for the other tab links. This will become especially useful to make sure the `Settings Updated` message is not shown on the other plugin setting tabs (when that update message is added into the core).