Class thead class name changes.
It may be to late, but I would say the class names for the `edd_cart_item_name` and `edd_cart_item_price` can be the same as the `tbody` names. If a user wants to target the `thead` area, they can simply use `.edd_cart_header_row`. The new method also keeps the naming convention with `edd_cart_actions` in both `thead` and `tbody`.