Add error message for a guest user at checkout
If "Require that users be logged-in to purchase files." is enabled, and "Display the registration and login forms on the checkout page for non-logged-in users." is disabled, there is no visual indication of what is happening (page just refreshes). This will fire an error message to let the user know that only logged in users can purchase files.