Add filter to item before template
It isn't currently possible (as far as I'm aware) to filter an item before it gets sent to the template (using Ajax).

It can be filtered using "edd_cart_item" but only *after* the template has been set, an item cannot be modified before this. Would be useful is custom options have been set in the item.