Fix active button state.
When using Twenty Twelve, the purchase button loses its border in the active state. This causes a jarring visual effect. Twenty Twelve adds `border: none` to the `input[type="submit"]:active` selector which causes the problem. Explicitly adding a border to the active state of the EDD purchase button fixes the issue.

Although this is EDD conflicting with a specific theme, I think it is an important change as it will help fix conflicts with Twenty Twelve child themes and make the button styling more stable across a number of themes.