Update template.php
If you use network_site_url() then it defaults to your primary site on the network, which is fine ... if you happen to be running from there. You may have the store as a subdomain, which in my case meant the images loaded from ipstenu.org while my SSL was setup on store.halfelf.org

Using site_url will fall back correctly and show the path to the site the page is on. And if you don't have SSL for THAT you have another problem :)