When there are multiple y axes, color code.
The Flot API is interesting - there are some open tickets on different areas to expand the ability to format axis labels, but as of right now, this semi-hacky way seems it might be the best approach.  I've tested several multiple axis reports, it seems like it should work fine.  But I'm über-open to there being a better way.