* Limit payments count by start and end date
Start and end date filters used on the payments table will now impact
the total count of payments shown in the views (status) and pagination
UIs on the table. Previously, the total number of payments was returned,
which meant that more results would be indicated in the views and
pagination than were actually displayed in the table.