Fix for when there is no $post on the page
This gives support for Pods Pages which can be created as standalone URLs apart from Post archives and singular templates. In certain cases, $post is not setup (like a 404 would have) and this throws errors when trying to utilize edd_get_purchase_link and passing in $args to manually set download_id.