Update view-order-details.php
Updated to add support for additional columns in the purchase downloads metabox list and to add columns in the 'add to purchase' row.  The motivation for the addition is to be able to support a currency selection when adding a new payment.  If additional columns are added, the plugin or theme css will need to implement styles for the expanded column list.