Added filter to allow modifying the variable prices
For simple downloads, the download price can be altered by adding a `edd_get_download_price` hook. The new hook will allow to do the same with variable prices, which, at the moment, cannot be modified if needed, because they are used straight away, in their raw form, as soon as they are loaded.