Allow downloads shortcode to show all available downloads
When the `posts_per_page` argument is set to `-1` in a WP_Query object, it will [return all posts](http://codex.wordpress.org/Class_Reference/WP_Query#Pagination_Parameters). However, in the downloads shortcode, this value is not possible, because the `number` variable is being sanitized with `absint()`, forcing a positive integer.

This change is valuable in cases where you don't know how many downloads will be displayed with the other given criteria, but you want to make sure you get all of them.