Add cart quantity to response for ajax remove
In order to display the actual correct amount of items in the cart, we need to pass it from PHP to JS using ajax - rather than just substracting using js. This allows us to show the correct amount of items in the cart even if an error happens to occur.