#603 - Initial Pass at Signed URLs
This is my initial pass. It's generating URLs with the new pattern, while keeping the old args in place for back compat for now.

I've verified that new urls and old urls (that were already generated) are downloading.

I've also verified that Old urls that are expired, or invalid don't work, as well as new urls with expirations or bad tokens fail to download.

I would love some feedback, testing, etc. Not counting on this being the only take. We have to be super careful not to mess this up.