Fix Ajax Test Works Function
Right now, we save whether or not the ajax test works as a transient, but we never actually use it. So what happens is on every edd_notices call, which is done on every admin page, EDD related or not, we're running this (can be) performance harming remote_post. This pr uses the transient to return early, so we only check it once a day, not every single notices call