Switch from $_SERVER["HTTPS"] check to is_ssl(). #3740.
is_ssl() handles the same check and an additional port check for when servers fail to set the HTTPS var. This removes the check from edd_enforced_ssl_redirect_handler() since it's already checked with is_ssl above, and it swaps the check in edd_get_current_page_url() to is_ssl().