Widgets admin fix for huge sites
By default, the "Download Details" widget form in the admin area attempts to load every download item into a drop-down. On sites with tens or hundreds of thousands of downloads, this causes wp-admin/widgets.php to break. This proposed change limits the drop-down to 1000 items, as opposed to -1 (no limit).