Store api_request in a transient for 24 hours.
We really only need to check through this hook once every 24 hours because on the updates page it checks every time.

I believe this is really just a backup check. If we don't restrict this check to every 24 hours, we end up with 2 checks for the exact same thing on every updates/plugin page load which slows down a lot of stuff.