Hook once to pre_set_site_transient_update_plugins
The "pre_set_site_transient_update_plugins" hook fires twice but we only need to use it the second time it fires. See trac ticket about this here: https://core.trac.wordpress.org/ticket/25542

This uses a boolean flag to let us know whether we are on the first time or the second time.