Trim states response
I was having an issue (as described here: https://github.com/easydigitaldownloads/easy-digital-downloads/issues/2963) where the states field was just showing `nostates` instead of the field.

Turns out the response from the ajax was: `     nostates`. this meant the conditional was not met. Added $.trim() to fix the issue and the field shows as it should.

Not sure why there were spaces before the response, but this solution works well and should avoid it in future.