Added filters to process the earnings as they are updated
The filter have the same purpose of the action added for customers, i.e. to allow processing the earnings before they are update on the product. This needed in cases like multi-currency shops, where earnings cannot just be added or subtracted, but must be first converted to the base currency.