Primary meta box base class
This commit adds the primary EDD_Meta_Box_Base class, as well as a postbox toggle in admin-scripts.js to enable toggling of EDD meta boxes.

class-meta-box-base.php is not required or included in this commit, and will likely be added in a later commit once it's determined where EDD core meta boxes are desired.

Implements #4781.