Allow requiring payment info on 100% discounted carts
Closes #5686

Adds a setting which, when true, will result in payment information
being required on checkout regardless of the value of a cart. This is
specifically to allow requiring payment info for a recurring subscription
when a coupon for 100% off the first billing period has been applied. Before
this patch, that situation would result in the buyer having to revisit
the site later on to provide payment information for their subscription to be
able to continue after the first billing period.