Add text to clarify when discount codes expire
Without this text, the time a discount code expires is rather ambiguous.
If you choose a date with the datepicker, you don't see until coming
back to edit the discount code that in fact a date and time (the last second of the
day) have been saved. This new text explicitly states that the discount
code will expire at the end of the day chosen.