First pass including custom database tables.
This change adds a new method to the primary singleton class for invoking all of the classes responsible for our custom database tables. As a first pass, it works well and the database tables are created correctly. Next, we will want to scrutinize and confirm method scope/name/approach/etc...

See #6277, #6308.