Introduce magic base class for querying any custom database table.
Adjust `EDD_DB_Order_Query` to be a subclass of this new class, and define the database columns in a new helper class: `EDD_DB_Column`.

`EDD_DB_Column` might be temporary, or not really necessary, but it did help me structure and scaffold much of the automation that's happening now.

See #6331.