Minimum PHP & WordPress Requirements Checker.
This change moves the main `Easy_Digital_Downloads` class into `includes/class-edd.php` and replaces the main class with a requirements checker.

This loads the text domain early (but not any other constants, classes, or functions) so that admin area requirements text can still be translated by the plugin correctly, even though the plugin itself is not fully loaded.

See #6290.