Introduce `failed()` method, and use it after item queries.
Also adds a `last_query` property, used to store the `WP_Error` if one was generated. This helps silence debug notices during regular user actiivty.

Also do not call `update_item` inside of `add_item` when the primary key is present. Methods are single-use, and it's not natural to magically try to update on add.