Allow empty strings and zero as meta values in Downloads.
Also add the `_edd_refund_window` download meta key.

This fixes a very old and elusive bug that defies the logic set forth by the setting labels, that an empty value would inherit the global, and 0 meant unlimited. In reality, it was never possible to leave this field empty, and it would erroneously report `0` even when the value was an empty string.

See #6652.