Introduce `reduce()` method to `Base` query class.
This is used to compare the current user's capabilites to actions being performed on item data.

For example, if a user cannot `update` a certain column, it will be stripped from the save data; if a user cannot `select` a certain column, it will be `unset()` (or set to `null` if it's already an object.)

Fixes #6508.