Merge the customer database table upgrades into 1 method
This change is a bit of a kludge. Some tables may already have the `status` column, and some may not.

These upgrade methods may show some noise in error logs, but that's usually only because the query was already altered but the upgrader is being ran anyways. In these cases, there's no good way to avoid doing those queries, and it also doesn't hurt to try - they just won't do anything.