Database: first pass UUID support for all object tables.
This change adds upgrade routines for all primary object database tables to add a `uuid` column to them. It also bumps those database table versions, and renames `generate_uuid()` to `validate_uuid()` and avoids updating the value on item-update calls.

Unit tests continue to pass, but also continue to be noisy from how database upgrade routines work in conjunction with how our unit tests are written to work. This can be fixed, but not as part of this. #yolofriday