Database: add `column_exists()` method to Base table class.
Use this new method to avoid adding columns that already exist, and attempting to modify columns that no longer exist.

This also changes the Base class to skip the upgrade routine if:

* Table exists
* Option does not exist

This broken condition is just skipped, which avoids running all upgrade routines, also avoiding database errors and debug notices in the process. Once beta is over, we can unplug this part - it will be nice to see these notices eventually.