Customers: updates to `EDD_Customer` and related.
This change soft-deprecates the increase/decrease customer methods, in favor of `recalculate_stats()`, which quickly counts and sums purchases.

It also tweaks the methods used to add/remove/make-primary email addresses, so that `customers:email` is a copy of the `primary` type in the normalized table.

Related is some left over minification of CSS files, and hopefully not any stray var_dump's.