Fields: ensure that all non-empty checkbox values compare correctly in `checked()`.
Positive checkbox values might be `on` or `''` or `1`, etc... In order for our multi-purpose HTML form-field renderer to correctly assess the checked value, we need to sanitize the values being compared against.

This change uses an `empty()` check to ensure that any falsey values are interpreted as unchecked, and truthy values are intepreted as checked.

Fixes #6912.