Dates: when presented with only a `gmt_offset` value, use math.
This change removes a frequently repeated and costly set of nested loops, that would search the pecl timezonedb looking for an offset match.

It also removes a `date( 'I' )` check, which attempted to check for daylight savings time in order to intelligently adjust the offset to search for. Unfortunately, WordPress always uses `UTC` as the default server timezone, which never sets DST, so this check would always fail, even if the site was in a timezone that recognized it.

This fixes a fatal error and a mathematical error when certain timezones were set in WordPress General Settings. (Pacific/Auckland, for example, is UTC+12 + 1 DST, but would match UTC+11:30. This no longer occurs.)

See #7016.