Support passing a primary column value in `add_item()`.
This allows for adding items with an accompanying auto-incremented `id` column, helping retain consistent ID data when migrating between tables.

When a primary column is identified, this performs an additional `get_item_raw()` call to look for an existing row with that value, and returns `false` immediately if one is found to avoid further MySQL errors, duplicate entries, or accidental updates.

See #7210.