Introduce `set_auto_increment()` to the databale `Table` class.
This utility method allows developers to manually set (or reset) the auto-increment value of a database table.

If a negative number or non-numeric value is passed in, it will default to `1` which is the lowest expected value.

MySQL is smart enough to not allow the database table to have an auto-increment value lower than the highest known value in that table. Attempting to set the value lower will result in the auto-increment value being set to the next highest number, without throwing errors or notices.