Add action before purchase receipt table (#7474)
* Add action before purchase receipt table

Lacking an alternative, I've seen several plugins (Jilt, Recurring Payments) (ab)use the `edd_payment_receipt_before` action to insert information before the table. For proper HTML this would need to be in a table row (the action bein in the `<thead` element), but layout wise this is not always desirable. Both the aforementioned plugins output a `div` and while the browser will do a best attempt to fix it, a simple additional action would be much better.

* edd_payment_receipt_before_table action hook documentation

* update edd_payment_receipt_before_table action hook documentation

Co-Authored-By: Spencer Finnell <spencerfinnell@gmail.com>

Co-authored-by: Spencer Finnell <spencerfinnell@gmail.com>