Update vertical tab roles and classes
Depending on whether JavaScript is being used for vertical sections, the
technical role of the section navigation and links are different:
- if JavaScript is being used, all content is loaded, and the sections
are really behaving more accurately as tabs
- if JavaScript is not being used (reports), the section links are most
accurately described as a menu

aria-selected is not a valid ARIA attribute for a menu item, but it is
for a tab. Additional ARIA markup helps clarify the correct role of the
list/items.

aria-current is a valid attribute for a menu item indicating the link is
for the current page.

This commit:
- updates the roles for the section-nav and section-title accordingly
- implements a new section-title--is-active CSS class to indicate the
current menu item/active tab
- changes the CSS to use the new class instead of the ARIA attribute
(since there are now two)

Fixes #7726