Create get_purchase_count and get_purchase_value customer methods
This separates the main work of the recalculate_stats method into two
methods, one for getting the count of orders and one for getting the
lifetime (purchase) value of the customer.

Recalculation is only performed within the recalculate_stats() method,
so the number displayed by the separate functions could be different
than what's in the database table.

#7947