Get the download's type with appropriate method
EDD_Download->__construct() doesn't set up the
download's type in EDD_Download->type. It is only
set up when get_type() is first called.

So, we need to make sure that we try to retrieve
the download's type by using the appropriate method,
instead of getting the $type property directly.