use table alias/name when building date query
use the received table name & table alias arguments in the `get_sql` function to set those as properties on the Date class.

Then, when building the SQL query, use the alias or name (only if available) to prepend to the column name in the query.

This prevents and fixes the issue described in https://github.com/awesomemotive/easy-digital-downloads/issues/9699